+++ whapps/voip/conference/conference.js
@@ -341,7 +341,9 @@ winkstart.module('voip', 'conference', {
             $('.conference-delete', conference_html).click(function(ev) {
                 ev.preventDefault();
 
-                THIS.delete_conference(data, callbacks.delete_success, callbacks.delete_error);
+                winkstart.confirm('Are you sure you want to delete this conference?', function() {
+                    THIS.delete_conference(data, callbacks.delete_success, callbacks.delete_error);
+                });
             });
 
             (target)
