+++ whapps/voip/media/media.js
@@ -299,7 +299,9 @@ winkstart.module('voip', 'media', {
             $('.media-delete', media_html).click(function(ev) {
                 ev.preventDefault();
 
-                THIS.delete_media(data, callbacks.delete_success, callbacks.delete_error);
+                winkstart.confirm('Are you sure you want to delete this media?', function() {
+                    THIS.delete_media(data, callbacks.delete_success, callbacks.delete_error);
+                });
             });
 
             (target)
