+++ whapps/voip/vmbox/vmbox.js
@@ -396,7 +396,9 @@ winkstart.module('voip', 'vmbox', {
             $('.vmbox-delete', vmbox_html).click(function(ev) {
                 ev.preventDefault();
 
-                THIS.delete_vmbox(data, callbacks.delete_success, callbacks.delete_error);
+                winkstart.confirm('Are you sure you want to delete this voicemail box?', function() {
+                    THIS.delete_vmbox(data, callbacks.delete_success, callbacks.delete_error);
+                });
             });
 
             (target)
